% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cloudml_create_config}
\alias{cloudml_create_config}
\title{Creates Default Configuration File}
\usage{
cloudml_create_config()
}
\description{
Creates a default 'cloudml.yml' configuration file to simplify the
configuration process. This file is created in the current directory,
open this file to configure this further.
}
