% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cloudml_deploy}
\alias{cloudml_deploy}
\title{Deploy SavedModel to CloudML}
\usage{
cloudml_deploy(export_dir_base)
}
\arguments{
\item{export_dir_base}{A string containing a directory containing an
exported SavedModels. Consider using \code{tensorflow::export_savedmodel()}
to export this SavedModel.}
}
\description{
Deploys a SavedModel to CloudML for online predictions.
}
