% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cloudml_init}
\alias{cloudml_init}
\title{Initializes Default Configuration File}
\usage{
cloudml_init()
}
\description{
Creates a default 'cloudml.yml' configuration file to initialize the
configuration process. This file is created in the current directory,
open this file to configure this further.
}
