% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud_local.R
\name{cloudml_local_predict}
\alias{cloudml_local_predict}
\title{Predict a Model Locally}
\usage{
cloudml_local_predict(model.dir = getwd(), json.instances = NULL,
  text.instances = NULL)
}
\arguments{
\item{model.dir}{The model directory.}

\item{json.instances}{Path to a JSON file, defining data
to be used for prediction.}

\item{text.instances}{Path to a text file, defining data
to be used for prediction.

TODO}
}
\description{
Predict a Model Locally
}
