% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{cloudml_local_train}
\alias{cloudml_local_train}
\title{Train a Model Locally}
\usage{
cloudml_local_train(application = getwd(),
  entrypoint = file.path(application, "app.R"), task.arguments = list())
}
\arguments{
\item{application}{The path to a TensorFlow application.}

\item{entrypoint}{The \R script which acts as an entrypoint to your
TensorFlow application.}

\item{arguments}{Additional arguments to be passed to your entrypoint.}
}
\description{
Train a Model Locally
}
