% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cloudml_predict}
\alias{cloudml_predict}
\title{Perform Prediction over a CloudML Model.}
\usage{
cloudml_predict(input, name = NULL, version = NULL, gcloud = NULL)
}
\arguments{
\item{name}{The name for this model. Defaults to the current directory
name.}

\item{version}{The version for this model. Versions start with a letter and
contain only letters, numbers and underscores. Defaults to the current
directory name.}
}
\description{
Perform online prediction over a CloudML model, usually, created using
\code{cloudml_deploy}.
}
\seealso{
\code{\link[=cloudml_deploy]{cloudml_deploy()}}
}
