% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{cloudml_tune}
\alias{cloudml_tune}
\title{Tune hyperparameters on Cloud ML}
\usage{
cloudml_tune(file = "train.R", config = "cloudml", scale_tier = c("basic",
  "basic-gpu", "basic-tpu"), flags = NULL, hypertune = "hypertune.yml")
}
\arguments{
\item{file}{File to be used as entrypoint for training.}

\item{config}{The name of the configuration to be used. Defaults to the
\code{"cloudml"} configuration.}

\item{scale_tier}{Machine type for training. "basic" provides a single
worker instance suitable for learning how to use Cloud ML Engine,
and for experimenting with new models using small datasets; "basic-gpu"
provides a single worker instance with a GPU. "basic-tpu" provides a
single worker instance with a TPU.}

\item{flags}{Named list with flag values (see \code{\link[=flags]{flags()}}) or path
to YAML file containing flag values.}

\item{hypertune}{Path to a YAML file, defining how hyperparameters should
be tuned. See
https://cloud.google.com/ml/reference/rest/v1/projects.jobs
for more details.}
}
\description{
Tune hyperparameters on Cloud ML
}
\seealso{
\code{\link[=job_trials]{job_trials()}}
}
