% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Get the currently active CloudML configuration}
\usage{
config()
}
\value{
R list with active configuration values.
}
\description{
Get the currently active CloudML configuration
}
\details{
The \code{config.yml} contains various configuration values. These values
can be distinct depending on whether execution is taking place on Google
Cloud or in a local context.

When running on Google Cloud the name of the configuration will always be
"gcloud". When running locally it will be "default" unless another
named configuration is made active.
}
\seealso{
\code{\link[config:get]{config::get()}}.
}
