% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud-config.R
\name{gcloud_account}
\alias{gcloud_account}
\title{Google Cloud Account}
\usage{
gcloud_account(account = NULL)
}
\arguments{
\item{account}{The account to use in subsequent Google Cloud SDK operations.
Usually the email associated to the account.}
}
\value{
The account associated to the Google Cloud SDK.
}
\description{
Gets/Sets the Google Cloud account through the Google Cloud SDK.
}
