% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud-config.R
\name{gcloud_config}
\alias{gcloud_config}
\title{Google Cloud Config}
\usage{
gcloud_config(gcloud = NULL)
}
\arguments{
\item{gcloud}{A list or \code{YAML} file with optional 'account', 'project',
and 'configuration' fields used to configure the GCloud environemnt.}
}
\description{
Reads the Google Cloud config file.
}
