% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud-install.R
\name{gcloud_install}
\alias{gcloud_install}
\title{Install the Google Cloud SDK}
\usage{
gcloud_install(version = "180.0.1")
}
\arguments{
\item{version}{Version of the Google Cloud SDK to be installed.}
}
\description{
Installs the Google Cloud SDK which enables CloudML operations.
}
