% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud-config.R
\name{gcloud_project}
\alias{gcloud_project}
\title{Google Cloud Project}
\usage{
gcloud_project(account, project)
}
\arguments{
\item{project}{The project to use in subsequent Google Cloud SDK operations.}
}
\value{
The project associated to the Google Cloud SDK.
}
\description{
Gets/Sets the Google Cloud account through the Google Cloud SDK.
}
