% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{gcloud_terminal}
\alias{gcloud_terminal}
\title{Create an RStudio terminal with access to the Google Cloud SDK}
\usage{
gcloud_terminal(command = NULL)
}
\arguments{
\item{command}{Command to send to terminal}
}
\value{
Terminal id (invisibly)
}
\description{
Create an RStudio terminal with access to the Google Cloud SDK
}
