% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud-version.R
\name{gcloud_version}
\alias{gcloud_version}
\title{Gcloud version}
\usage{
gcloud_version()
}
\value{
a list with the version of each component.
}
\description{
Get version of Google Cloud SDK components.
}
