% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud.R
\name{gs_copy}
\alias{gs_copy}
\title{Copy files from Google Storage}
\usage{
gs_copy(uri, destination, overwrite = FALSE)
}
\arguments{
\item{uri}{Google storage URI (e.g. \code{gs://[BUCKET_NAME]/[FILENAME.CSV]})}

\item{destination}{Path to copy file to on the local filesystem}

\item{overwrite}{Overwrite an existing file of the same name}
}
\description{
Copy files from Google Storage
}
