% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud-storage.R
\name{gs_local_dir}
\alias{gs_local_dir}
\title{Alias to gs_data_dir_local() function}
\usage{
gs_local_dir(url, local_dir = "gs", echo = FALSE)
}
\arguments{
\item{url}{Google Storage bucket URL (e.g. \code{gs://<your-bucket>}).}

\item{local_dir}{Local directory to synchonize Google Storage bucket(s) to.}

\item{echo}{Echo command output to console.}
}
\description{
This function is deprecated, please use \code{\link[=gs_data_dir_local]{gs_data_dir_local()}} instead.
}
\seealso{
\code{\link[=gs_data_dir_local]{gs_data_dir_local()}}
}
\keyword{internal}
