% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsutil-copy.R
\name{gsutil_data}
\alias{gsutil_data}
\title{Get local path to data within Google Storage}
\usage{
gsutil_data(uri, local_dir = "gs")
}
\arguments{
\item{uri}{Path to Google Storage data}

\item{local_dir}{Local directory to copy files into}
}
\value{
Path to data file (may be local or remote depending on the execution
context).
}
\description{
When running on Google Cloud the path is returned unchanged. When running in
other contexts the file is copied to the local system and a path to the local
file is returned. If a plain filesystem path is passed then it is also
returned unchanged.
}
