% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsutil-exec.R
\name{gsutil_exec}
\alias{gsutil_exec}
\title{Executes a Google Utils Command}
\usage{
gsutil_exec(..., args = NULL)
}
\arguments{
\item{...}{Parameters to use specified based on position.}

\item{args}{Parameters to use specified as a list.}
}
\description{
Executes a Google Utils command with the given parameters.
}
