% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud-storage.R
\name{is_cloudml}
\alias{is_cloudml}
\title{Detect when executing on CloudML}
\usage{
is_cloudml()
}
\value{
\code{TRUE} when the current script is executing on CloudML, otherwise \code{FALSE}
}
\description{
Detect when executing on CloudML
}
