% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job-utils.R
\name{job_dir}
\alias{job_dir}
\title{Generate a Job Directory}
\usage{
job_dir(prefix = "jobs")
}
\arguments{
\item{prefix}{The prefix to be used for the job directory.}
}
\description{
Generate a job directory (as a relative path). Useful for
deployments when you want model artefacts to be confined to
a unique directory.
}
