% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_describe}
\alias{job_describe}
\alias{job_stream}
\alias{job_status}
\title{Get job information}
\usage{
job_describe(job)

job_stream(job, polling_interval = 60, task_name = NULL,
  allow_multiline_logs = FALSE)

job_status(job)
}
\arguments{
\item{job}{Either a `cloudml_job` object as returned by [cloudml_train()],
or the name of a job.}

\item{polling_interval}{Polling interval for streamed output.}

\item{task_name}{If set, display only the logs for this particular task.}

\item{allow_multiline_logs}{Output multiline log messages as single records.}
}
\description{
Get detailed information on a job and it's status. Stream
the log of a running job.
}
\seealso{
Other jobs: \code{\link{cloudml_train}},
  \code{\link{job_cancel}}, \code{\link{job_collect}},
  \code{\link{job_list}}
}
