% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_stream}
\alias{job_stream}
\title{Google Cloud -- Stream Logs from a Job}
\usage{
job_stream(job, polling_interval = 60, task_name = NULL,
  allow_multiline_logs = FALSE)
}
\arguments{
\item{job}{Either a `cloudml_job` object as returned by [cloudml_train()],
or the name of a job.}

\item{polling_interval}{Number of seconds to wait between efforts to fetch the
latest log messages.}

\item{task_name}{If set, display only the logs for this particular task.}

\item{allow_multiline_logs}{Output multiline log messages as single records.}
}
\description{
Show logs from a running Cloud ML Engine job.
}
\seealso{
Other jobs: \code{\link{cloudml_train}},
  \code{\link{job_cancel}}, \code{\link{job_collect}},
  \code{\link{job_describe}}, \code{\link{job_list}},
  \code{\link{job_status}}
}
