% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_trials}
\alias{job_trials}
\title{Current trials of a job}
\usage{
job_trials(x)
}
\arguments{
\item{x}{Job name or job object.}
}
\description{
Get the hyperparameter trials for job, as an \R data frame
}
\seealso{
Other job management: \code{\link{job_cancel}},
  \code{\link{job_collect_async}},
  \code{\link{job_collect}}, \code{\link{job_list}},
  \code{\link{job_status}}, \code{\link{job_stream_logs}}
}
\concept{job management}
