% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{local_predict}
\alias{local_predict}
\title{Generate Predictions with a Model Locally}
\usage{
local_predict(model_dir, data)
}
\arguments{
\item{model_dir}{The path to a model directory, or a jobs directory containing
an exported model directory.}

\item{data}{The dataset to be used for prediction.}
}
\description{
Generate predictions using a TensorFlow model saved on disk.
}
