% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R, R/predict.R
\name{predict_local}
\alias{predict_local}
\alias{predict_local}
\title{Predict a Model Locally}
\usage{
predict_local(dir, data)

predict_local(dir, data)
}
\arguments{
\item{dir}{The path to a model directory, or a jobs directory containing
an exported model directory.}

\item{data}{The dataset to be used for prediction.}

\item{model.dir}{The model directory.}

\item{json.instances}{Path to a JSON file, defining data
to be used for prediction.}

\item{text.instances}{Path to a text file, defining data
to be used for prediction.

TODO}
}
\description{
Predict a Model Locally

Generate predictions using a TensorFlow model saved on disk.
}
