% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{sdk_command}
\alias{sdk_command}
\title{Trigger Google Cloud SDK Command}
\usage{
sdk_command(...)
}
\arguments{
\item{...}{Parameters to use with 'gcloud'.}
}
\description{
Triggers a Google Cloud SDK command.
}
