% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdk.R
\name{sdk_config}
\alias{sdk_config}
\title{Config the Google Cloud SDK}
\usage{
sdk_config(account, project)
}
\arguments{
\item{account}{The account to use in subsequent Google Cloud SDK operations.
Usually the email associated to the account.}

\item{project}{The project to use in subsequent Google Cloud SDK operations.}
}
\description{
Config the Google Cloud SDK to enable CloudML operations.
}
