% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{train_cloudml}
\alias{train_cloudml}
\title{Google Cloud -- Submit a Training Job}
\usage{
train_cloudml(application = getwd(), config = "cloudml", ...)
}
\arguments{
\item{application}{The path to a TensorFlow application.}

\item{config}{The name of the configuration to be used.}

\item{...}{Optional arguments. TODO: describe in more detail.}
}
\description{
Upload a TensorFlow application to Google Cloud, and use that application
to train a model.
}
