% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job-utils.R
\name{unique_job_dir}
\alias{unique_job_dir}
\title{Generate a Unique Job Directory}
\usage{
unique_job_dir(prefix = "")
}
\arguments{
\item{prefix}{The prefix to be used for the job directory.}
}
\description{
Generate a job directory (as a relative path). Useful for
deployments when you want model artefacts to be confined to
a unique directory.
}
