% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudml-package.R
\docType{package}
\name{cloudml-package}
\alias{cloudml}
\alias{cloudml-package}
\title{R Interface to Google Cloud ML}
\description{
This package provides an interface to Google's Cloud Machine Learning
platform.
}
\author{
\strong{Maintainer}: JJ Allaire \email{jj@rstudio.com}

Authors:
\itemize{
  \item Kevin Ushey \email{kevin@rstudio.com}
}

}
