% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cloudml_train}
\alias{cloudml_train}
\title{Train a model using Cloud ML}
\usage{
cloudml_train(file = "train.R", master_type = NULL, flags = NULL,
  region = NULL, config = NULL, collect = "ask")
}
\arguments{
\item{file}{File to be used as entrypoint for training.}

\item{master_type}{Training master node machine type. "standard" provides a
basic machine configuration suitable for training simple models with small
to moderate datasets. See the documentation at
https://cloud.google.com/ml-engine/docs/training-overview#machine_type_table
for details on available machine types.}

\item{flags}{Named list with flag values (see \code{\link[=flags]{flags()}}) or path
to YAML file containing flag values.}

\item{region}{The region to be used for training.}

\item{config}{A list, \code{YAML} or \code{JSON} configuration file as described
https://cloud.google.com/ml-engine/reference/rest/v1/projects.jobs.}

\item{collect}{Collect job when training is completed (blocks waiting for the
job to complete).}
}
\description{
Upload a TensorFlow application to Google Cloud, and use that application to
train a model.
}
\seealso{
\code{\link[=job_status]{job_status()}}, \code{\link[=job_collect]{job_collect()}}, \code{\link[=job_cancel]{job_cancel()}}

Other CloudML functions: \code{\link{cloudml_deploy}},
  \code{\link{cloudml_predict}}
}
