% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cloudml_train}
\alias{cloudml_train}
\title{Train a model using Cloud ML}
\usage{
cloudml_train(file = "train.R", config = "cloudml",
  scale_tier = c("basic", "basic-gpu", "basic-tpu"), flags = NULL,
  hypertune = NULL, collect = "ask")
}
\arguments{
\item{file}{File to be used as entrypoint for training.}

\item{config}{The name of the configuration to be used. Defaults to the
\code{"cloudml"} configuration.}

\item{scale_tier}{Machine type for training. "basic" provides a single
worker instance suitable for learning how to use Cloud ML Engine,
and for experimenting with new models using small datasets; "basic-gpu"
provides a single worker instance with a GPU. "basic-tpu" provides a
single worker instance with a TPU.}

\item{flags}{Named list with flag values (see \code{\link[=flags]{flags()}}) or path
to YAML file containing flag values.}

\item{hypertune}{Path to a YAML file, defining how hyperparameters should
be tuned. See
https://cloud.google.com/ml/reference/rest/v1/projects.jobs
for more details.}

\item{collect}{Collect job output after submission. Defaults to "ask" which
will prompt within interactive environments. Within versions of RStudio
that support terminals (>= v1.1), the waiting and collection will be done
within a terminal.}

\item{...}{Named arguments, used to supply runtime configuration settings to
your TensorFlow application.}
}
\description{
Upload a TensorFlow application to Google Cloud, and use that application to
train a model.
}
\seealso{
\code{\link[=job_describe]{job_describe()}}, \code{\link[=job_collect]{job_collect()}}, \code{\link[=job_cancel]{job_cancel()}}
}
