% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cloudml_train}
\alias{cloudml_train}
\title{Train a model using Cloud ML}
\usage{
cloudml_train(application = getwd(), config = "cloudml", ...)
}
\arguments{
\item{application}{The path to a TensorFlow application. Defaults to
the current working directory.}

\item{config}{The name of the configuration to be used. Defaults to
the \code{"cloudml"} configuration.}

\item{...}{Named arguments, used to supply runtime configuration
settings to your TensorFlow application.}
}
\description{
Upload a TensorFlow application to Google Cloud, and use that application to
train a model.
}
\seealso{
\code{\link[=job_describe]{job_describe()}}, \code{\link[=job_collect]{job_collect()}}, \code{\link[=job_cancel]{job_cancel()}}
}
