% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{cloudml_train}
\alias{cloudml_train}
\title{Train a model using Cloud ML}
\usage{
cloudml_train(application = getwd(), config = "cloudml", job_dir = NULL,
  ...)
}
\arguments{
\item{application}{The path to a TensorFlow application}

\item{config}{The name of the configuration to be used.}

\item{job_dir}{Directory to write job into (defaults to the value
of \code{job_dir} in the \code{config.yml} file).}

\item{...}{Named arguments, used to supply runtime configuration
settings to your TensorFlow application. When
\code{\link[cloudml:config]{cloudml::config()}} is called, these values will effectively
be overlayed on top of the configuration requested.}
}
\description{
Upload a TensorFlow application to Google Cloud, and use that application to
train a model.
}
\seealso{
\code{\link[=job_describe]{job_describe()}}, \code{\link[=job_collect]{job_collect()}}, \code{\link[=job_cancel]{job_cancel()}}
}
