% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{cloudml_tune}
\alias{cloudml_tune}
\title{Tune hyperparameters on Cloud ML}
\usage{
cloudml_tune(application = getwd(), config = "cloudml",
  hyperparameters = "hyperparameters.yml", ...)
}
\arguments{
\item{application}{The path to a TensorFlow application. Defaults to
the current working directory.}

\item{config}{The name of the configuration to be used. Defaults to
the \code{"cloudml"} configuration.}

\item{hyperparameters}{Path to the job configuration file. The file should be
a YAML document containing a Job resource as defined in
the API (all fields are optional):
https://cloud.google.com/ml/reference/rest/v1/projects.jobs}

\item{...}{Named arguments, used to supply runtime configuration
settings to your TensorFlow application. When
\code{\link[cloudml:config]{cloudml::config()}} is called, these values will effectively
be overlayed on top of the configuration requested.}
}
\description{
Tune hyperparameters on Cloud ML
}
