% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Read the configuration for a CloudML application}
\usage{
config(config = "default", local_gs = "gs")
}
\arguments{
\item{config}{Configuration name (e.g. "default", "cloudml", etc.)}

\item{local_gs}{Specify a directory name to automatically download local
copies of references to Google Storage data (\code{gs://}) and then
resolve their values to the path of their local copy.}
}
\value{
List with configuration values
}
\description{
Read the configuration for a CloudML application
}
