% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Read the configuration for a Cloud ML application}
\usage{
config(config = NULL, local_gs = "local/gs")
}
\arguments{
\item{config}{Name of configuration to read (\code{NULL} will result in the
active configuration for the current environment being used).}

\item{local_gs}{Specify a directory name to automatically download local
copies of references to Google Storage data (\code{gs://}) and then
resolve their values to the path of their local copy.}
}
\value{
List with configuration values
}
\description{
Read the configuration for a Cloud ML application
}
