% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud-config.R
\name{gcloud_config}
\alias{gcloud_config}
\title{Google Cloud Config}
\usage{
gcloud_config(path = getwd())
}
\arguments{
\item{path}{Path to 'cloudml.yml' file; defaults to \code{getwd()}.}
}
\description{
Reads the Google Cloud config file.
}
