% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud-install.R
\name{gcloud_install}
\alias{gcloud_install}
\title{Install the Google Cloud SDK}
\usage{
gcloud_install(version = "178.0.0")
}
\description{
Installs the Google Cloud SDK which enables CloudML operations.
}
