% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{gcloud_terminal}
\alias{gcloud_terminal}
\title{Create an RStudio terminal with access to the Google Cloud SDK}
\usage{
gcloud_terminal(command = NULL, clear = FALSE)
}
\arguments{
\item{command}{Command to send to terminal}

\item{clear}{Clear terminal buffer}
}
\value{
Terminal id (invisibly)
}
\description{
Create an RStudio terminal with access to the Google Cloud SDK
}
\seealso{
Other Google Cloud SDK functions: \code{\link{gcloud_init}},
  \code{\link{gcloud_install}}
}
\concept{Google Cloud SDK functions}
