% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud.R
\name{gs_data}
\alias{gs_data}
\title{Get a path to data within Google Storage}
\usage{
gs_data(uri, local_dir = "local/gs")
}
\arguments{
\item{uri}{Google storage URI (e.g. \code{gs://[BUCKET_NAME]/[FILENAME.CSV]})}

\item{local_dir}{Local directory to copy files into}
}
\value{
Path to data file (may be local or remote depending on the execution
context).
}
\description{
When running on Google Cloud the path is returned unchanged. When running in
other contexts the file is copied to the local system and a path to the local
file is returned. If a plain filesystem path is passed then it is also
returned unchanged.
}
