% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_cancel}
\alias{job_cancel}
\title{Cancel a job}
\usage{
job_cancel(job)
}
\arguments{
\item{job}{Either a \code{cloudml_job} object as returned by \code{\link[=cloudml_train]{cloudml_train()}},
or the name of a job.}
}
\description{
Cancel a job.
}
\seealso{
\code{\link[=job_describe]{job_describe()}}, \code{\link[=job_collect]{job_collect()}}, \code{\link[=job_list]{job_list()}}
}
