% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_cancel}
\alias{job_cancel}
\title{Google Cloud -- Cancel a Job}
\usage{
job_cancel(job)
}
\arguments{
\item{job}{Either a `cloudml_job` object as returned by [train_cloudml()],
or the name of a job.}
}
\description{
Cancel a job.
}
\seealso{
Other jobs: \code{\link{job_collect}},
  \code{\link{job_describe}}, \code{\link{job_list}},
  \code{\link{job_status}}, \code{\link{job_stream}},
  \code{\link{train_cloudml}}
}
