% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_cancel}
\alias{job_cancel}
\title{Cancel a job}
\usage{
job_cancel(job, gcloud = NULL)
}
\arguments{
\item{job}{Job name or job object.}

\item{gcloud}{A list or \code{YAML} file with optional 'account' or 'project'
fields used to configure the GCloud environemnt.}
}
\description{
Cancel a job.
}
\seealso{
Other job management: \code{\link{job_collect_async}},
  \code{\link{job_collect}}, \code{\link{job_list}},
  \code{\link{job_status}}, \code{\link{job_stream_logs}},
  \code{\link{job_trials}}
}
