% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_cancel}
\alias{job_cancel}
\title{Cancel a job}
\usage{
job_cancel(job = "latest")
}
\arguments{
\item{job}{Job name or job object. Pass "latest" to indicate the
most recently submitted job.}
}
\description{
Cancel a job.
}
\seealso{
Other job management functions: 
\code{\link{job_collect}()},
\code{\link{job_list}()},
\code{\link{job_status}()},
\code{\link{job_stream_logs}()},
\code{\link{job_trials}()}
}
\concept{job management functions}
