% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_cancel}
\alias{job_cancel}
\title{Cancel a job}
\usage{
job_cancel(job)
}
\arguments{
\item{job}{Job name or job object.}
}
\description{
Cancel a job.
}
\seealso{
Other job management: \code{\link{job_collect_async}},
  \code{\link{job_collect}}, \code{\link{job_describe}},
  \code{\link{job_list}}, \code{\link{job_status}},
  \code{\link{job_stream}}
}
\concept{job management}
