% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_collect}
\alias{job_collect}
\title{Collect job output}
\usage{
job_collect(job, destination = "jobs/cloudml")
}
\arguments{
\item{job}{Either a \code{cloudml_job} object as returned by \code{\link[=cloudml_train]{cloudml_train()}},
or the name of a job.}

\item{destination}{The destination directory in which model outputs should
be downloaded. Defaults to \code{jobs/cloudml}.}
}
\description{
Collect the job outputs (e.g. fitted model) from a job.
If the job has not yet finished running, \code{job_collect()}
will block and wait until the job has finished.
}
\seealso{
\code{\link[=job_describe]{job_describe()}}, \code{\link[=job_cancel]{job_cancel()}}, \code{\link[=job_list]{job_list()}}
}
