% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_collect}
\alias{job_collect}
\title{Collect Results from a Job}
\usage{
job_collect(job, destination = "jobs/cloudml")
}
\arguments{
\item{job}{Either a `cloudml_job` object as returned by [cloudml_train()],
or the name of a job.}

\item{destination}{The destination directory in which model outputs should
be downloaded. Defaults to \code{jobs/cloudml}.}
}
\description{
Collect the job outputs (e.g. fitted model) from a job.
If the job has not yet finished running, \code{job_collect()}
will block and wait until the job has finished.
}
\seealso{
Other jobs: \code{\link{cloudml_train}},
  \code{\link{job_cancel}}, \code{\link{job_describe}},
  \code{\link{job_list}}, \code{\link{job_status}},
  \code{\link{job_stream}}
}
