% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_describe}
\alias{job_describe}
\title{Google Cloud -- Describe a Job}
\usage{
job_describe(job)
}
\arguments{
\item{job}{Either a `cloudml_job` object as returned by [cloudml_train()],
or the name of a job.}
}
\description{
Describe a job.
}
\seealso{
Other jobs: \code{\link{cloudml_train}},
  \code{\link{job_cancel}}, \code{\link{job_collect}},
  \code{\link{job_list}}, \code{\link{job_status}},
  \code{\link{job_stream}}
}
