% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_describe}
\alias{job_describe}
\title{Describe a job}
\usage{
job_describe(job)
}
\arguments{
\item{job}{Job name or job object.}
}
\description{
Describe a job
}
\seealso{
Other job management: \code{\link{job_cancel}},
  \code{\link{job_collect}}, \code{\link{job_list}},
  \code{\link{job_status}}, \code{\link{job_stream}}
}
\concept{job management}
