% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_dir}
\alias{job_dir}
\title{Generate a Job Directory}
\usage{
job_dir(prefix = "jobs")
}
\description{
Generate a job directory (as a relative path). Useful for
deployments when you want model artefacts to be confined to
a unique directory.
}
