% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_status}
\alias{job_status}
\title{Current status of a job}
\usage{
job_status(job = "latest", gcloud = NULL)
}
\arguments{
\item{job}{Job name or job object. Pass "latest" to indicate the
most recently submitted job.}

\item{gcloud}{A list or \code{YAML} file with optional 'account' or 'project'
fields used to configure the GCloud environemnt.}
}
\description{
Get the status of a job, as an \R list.
}
\seealso{
Other job management: \code{\link{job_cancel}},
  \code{\link{job_collect}}, \code{\link{job_list}},
  \code{\link{job_stream_logs}}, \code{\link{job_trials}}
}
