% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_status}
\alias{job_status}
\title{Current status of a job}
\usage{
job_status(job)
}
\arguments{
\item{job}{Either a `cloudml_job` object as returned by [cloudml_train()],
or the name of a job.}
}
\description{
Get the status of a job, as an \R list.
}
\seealso{
Other jobs: \code{\link{cloudml_train}},
  \code{\link{job_cancel}}, \code{\link{job_collect}},
  \code{\link{job_describe}}, \code{\link{job_list}},
  \code{\link{job_stream}}
}
