% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_stream_logs}
\alias{job_stream_logs}
\title{Show job log stream}
\usage{
job_stream_logs(job,
  polling_interval = getOption("cloudml.stream_logs.polling", 5),
  task_name = NULL, allow_multiline_logs = FALSE, gcloud = NULL)
}
\arguments{
\item{job}{Job name or job object.}

\item{polling_interval}{Number of seconds to wait between efforts to fetch the
latest log messages.}

\item{task_name}{If set, display only the logs for this particular task.}

\item{allow_multiline_logs}{Output multiline log messages as single records.}

\item{gcloud}{A list or \code{YAML} file with optional 'account' or 'project'
fields used to configure the GCloud environemnt.}
}
\description{
Show logs from a running Cloud ML Engine job.
}
\seealso{
Other job management: \code{\link{job_cancel}},
  \code{\link{job_collect_async}},
  \code{\link{job_collect}}, \code{\link{job_list}},
  \code{\link{job_status}}, \code{\link{job_trials}}
}
