% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{local_train}
\alias{local_train}
\title{Train a model locally}
\usage{
local_train(application = getwd(), config = "default",
  entrypoint = "train.R", ...)
}
\arguments{
\item{application}{The path to a TensorFlow application. Defaults to
the current working directory.}

\item{config}{The name of the configuration to be used. Defaults to
the \code{"cloudml"} configuration.}

\item{...}{Named arguments, used to supply runtime configuration
settings to your TensorFlow application.}
}
\description{
Train a model locally, using the \code{gcloud} command line
utility. This can be used as a testing bed for TensorFlow
applications which you want to later run on Google Cloud,
submitted using \code{\link[=cloudml_train]{cloudml_train()}}.
}
