% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{train_cloud}
\alias{train_cloud}
\title{Google Cloud -- Submit a Training Job}
\usage{
train_cloud(application = getwd(), config = "gcloud", async = TRUE, ...)
}
\arguments{
\item{application}{The path to a TensorFlow application.}

\item{config}{The name of the configuration to be used.}

\item{async}{Run the job asynchronously? When `FALSE`, this call will
block until the TensorFlow job has run to completion.}

\item{...}{Additional configuration values for this training run}

\item{...}{Optional arguments. TODO: describe in more detail.}
}
\description{
Upload a TensorFlow application to Google Cloud, and use that application
to train a model.
}
